Move tx submission through transactor
